Moved all tikz related headers in directory sealibvisual. Moved the associated .cpp files into src-view/
