+++ src-view/main.cpp
@@ -1,8 +1,8 @@
-#include <sealib/tikznode.h>
-#include <sealib/tikzarray.h>
-#include <sealib/tikzdocument.h>
-#include <sealib/tikzpicture.h>
-#include <sealib/tikzgenerator.h>
+#include <include/sealibvisual/tikznode.h>
+#include <include/sealibvisual/tikzarray.h>
+#include <include/sealibvisual/tikzdocument.h>
+#include <include/sealibvisual/tikzpicture.h>
+#include <include/sealibvisual/tikzgenerator.h>
 #include <sealib/graphcreator.h>
 #include <sealib/dfs.h>
 #include <sealib/graphrepresentations.h>
